Introduce AwardType class constants to replace magic numbers
Move classes from `lib/` to individual files.
