+++ lib/database/forum.php
@@ -1,15 +1,7 @@
 <?php
 
+use RA\ModifyTopicField;
 use RA\Permissions;
-
-abstract class ModifyTopicField
-{
-    public const ModifyTitle = 0;
-
-    public const DeleteTopic = 1;
-
-    public const RequiredPermissions = 2;
-}
 
 function getForumList($categoryID = 0)
 {
