+++ lib/database/user.php
@@ -3,35 +3,6 @@
 use RA\ActivityType;
 use RA\Permissions;
 
-abstract class UserPref
-{
-    public const EmailOn_ActivityComment = 0;
-
-    public const EmailOn_AchievementComment = 1;
-
-    public const EmailOn_UserWallComment = 2;
-
-    public const EmailOn_ForumReply = 3;
-
-    public const EmailOn_AddFriend = 4;
-
-    public const EmailOn_PrivateMessage = 5;
-
-    public const EmailOn_Newsletter = 6;
-
-    public const EmailOn_unused2 = 7;
-
-    public const SiteMsgOn_ActivityComment = 8;
-
-    public const SiteMsgOn_AchievementComment = 9;
-
-    public const SiteMsgOn_UserWallComment = 10;
-
-    public const SiteMsgOn_ForumReply = 11;
-
-    public const SiteMsgOn_AddFriend = 12;
-}
-
 //////////////////////////////////////////////////////////////////////////////////////////
 //    Accounts
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -1509,22 +1480,22 @@ function GetDeveloperStatsFull($count, $sortBy, $devFilter = 7)
 
     switch ($devFilter) {
         case 1: // Active
-            $stateCond = " AND ua.Permissions >= " . \RA\Permissions::Developer;
+            $stateCond = " AND ua.Permissions >= " . Permissions::Developer;
             break;
         case 2: // Junior
-            $stateCond = " AND ua.Permissions = " . \RA\Permissions::JuniorDeveloper;
+            $stateCond = " AND ua.Permissions = " . Permissions::JuniorDeveloper;
             break;
         case 3: // Active + Junior
-            $stateCond = " AND ua.Permissions >= " . \RA\Permissions::JuniorDeveloper;
+            $stateCond = " AND ua.Permissions >= " . Permissions::JuniorDeveloper;
             break;
         case 4: // Inactive
-            $stateCond = " AND ua.Permissions <= " . \RA\Permissions::Registered;
+            $stateCond = " AND ua.Permissions <= " . Permissions::Registered;
             break;
         case 5: // Active + Inactive
-            $stateCond = " AND ua.Permissions <> " . \RA\Permissions::JuniorDeveloper;
+            $stateCond = " AND ua.Permissions <> " . Permissions::JuniorDeveloper;
             break;
         case 6: // Junior + Inactive
-            $stateCond = " AND ua.Permissions <= " . \RA\Permissions::JuniorDeveloper;
+            $stateCond = " AND ua.Permissions <= " . Permissions::JuniorDeveloper;
             break;
         case 0: // Active + Junior + Inactive
         case 7:
