+++ src/AwardType.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace RA;
+
+abstract class AwardType
+{
+    public const MASTERY = 1;
+
+    public const ACHIEVEMENT_UNLOCKS_YIELD = 2;
+
+    public const ACHIEVEMENT_POINTS_YIELD = 3;
+
+    public const REFERRALS = 4;
+
+    public const FACEBOOK_CONNECT = 5;
+
+    public const PATREON_SUPPORTER = 6;
+}
