+++ src/ModifyTopicField.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace RA;
+
+abstract class ModifyTopicField
+{
+    public const ModifyTitle = 0;
+
+    public const DeleteTopic = 1;
+
+    public const RequiredPermissions = 2;
+}
