+++ src/UserPref.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace RA;
+
+abstract class UserPref
+{
+    public const EmailOn_ActivityComment = 0;
+
+    public const EmailOn_AchievementComment = 1;
+
+    public const EmailOn_UserWallComment = 2;
+
+    public const EmailOn_ForumReply = 3;
+
+    public const EmailOn_AddFriend = 4;
+
+    public const EmailOn_PrivateMessage = 5;
+
+    public const EmailOn_Newsletter = 6;
+
+    public const EmailOn_unused2 = 7;
+
+    public const SiteMsgOn_ActivityComment = 8;
+
+    public const SiteMsgOn_AchievementComment = 9;
+
+    public const SiteMsgOn_UserWallComment = 10;
+
+    public const SiteMsgOn_ForumReply = 11;
+
+    public const SiteMsgOn_AddFriend = 12;
+}
